Refactored second task
