+++ lib/main.dart
@@ -1,7 +1,6 @@
-import 'package:first_project/models/user.dart';
-import 'package:first_project/providers/user_provider.dart';
+import 'package:first_project/providers/user_notifier.dart';
+import 'package:first_project/routes/app_routes.dart';
 import 'package:first_project/screens/home_page.dart';
-import 'package:first_project/screens/public_profile_page.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -15,20 +14,10 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ChangeNotifierProvider(
-      create: (_) => UserProvider(),
-      child: MaterialApp(
+      create: (_) => UserNotifier(),
+      child: const MaterialApp(
         initialRoute: HomePage.routeName,
-        onGenerateRoute: (settings) {
-          if (settings.name == PublicProfilePage.routeName) {
-            final user = settings.arguments as User;
-            return MaterialPageRoute(
-              builder: (context) => PublicProfilePage(user: user),
-            );
-          }
-          return MaterialPageRoute(
-            builder: (context) => const HomePage(),
-          );
-        },
+        onGenerateRoute: generateRoute,
       ),
     );
   }
