+++ lib/models/user.dart
@@ -1,4 +1,5 @@
 import 'package:first_project/enums/connection_status.dart';
+import 'package:first_project/enums/leave_type.dart';
 import 'package:first_project/enums/role.dart';
 import 'package:first_project/models/leave_request.dart';
 
@@ -7,22 +8,17 @@ class User {
   final String surname;
   final Role role;
   String profilePicture;
-  ConnectionStatus _status = ConnectionStatus.Offline;
+  ConnectionStatus status = ConnectionStatus.Offline;
   List<LeaveRequest> requests;
-  bool isAbsent = false;
+  LeaveType? currentLeaveType;
 
-  User(this.name, this.surname, this.role, this.profilePicture, this._status, {this.requests = const []});
+  User(this.name, this.surname, this.role, this.profilePicture, this.status, {this.requests = const []});
 
   void toggleConnectionStatus() {
-    _status = _status == ConnectionStatus.Offline ? ConnectionStatus.Online : ConnectionStatus.Online;
+    status = status == ConnectionStatus.Offline ? ConnectionStatus.Online : ConnectionStatus.Online;
   }
 
   void createRequest(LeaveRequest request) {
     requests.add(request);
-  }
-
-  ConnectionStatus get status => _status;
-  set status(ConnectionStatus newStatus) {
-    _status = newStatus;
   }
 }
