+++ lib/ui/user_tiles.dart
@@ -1,5 +1,5 @@
 import 'package:first_project/models/user.dart';
-import 'package:first_project/providers/user_provider.dart';
+import 'package:first_project/providers/user_notifier.dart';
 import 'package:first_project/ui/user_container.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -11,7 +11,7 @@ class UserTiles extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    List<User> users = context.watch<UserProvider>().users;
+    List<User> users = context.watch<UserNotifier>().users;
     return Expanded(
       child: ListView.builder(
         itemCount: users.length,
